requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
transformers[torch]
numpy
requests
sentencepiece
nltk
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp38-cp38-linux_x86_64.whl; sys_platform == 'linux' and python_version == '3.8'